UC10: Added a new class for dictionary services and a method to count the persons in a city
